+++ test/IbanNet.Tests/TypeConverters/IbanTypeConverterTests.cs
@@ -2,6 +2,7 @@
 using System.ComponentModel;
 using System.Globalization;
 using FluentAssertions;
+using IbanNet.Registry;
 using Moq;
 using Newtonsoft.Json;
 using TestHelpers;
@@ -94,7 +95,7 @@ namespace IbanNet.TypeConverters
 
             public When_converting_to_string()
             {
-                _iban = new Iban(TestValues.ValidIban);
+                _iban = new IbanParser(IbanRegistry.Default).Parse(TestValues.ValidIban);
             }
 
             [Fact]
@@ -136,7 +137,7 @@ namespace IbanNet.TypeConverters
             [Fact]
             public void It_should_succeed()
             {
-                var bankAccountNumber1 = new Iban(TestValues.ValidIban);
+                Iban bankAccountNumber1 = new IbanParser(IbanRegistry.Default).Parse(TestValues.ValidIban);
                 string json = JsonConvert.SerializeObject(bankAccountNumber1);
 
                 json.Should().Be($"\"{TestValues.ValidIban}\"");